pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-shell"
version = "0.1.0"
description = "Natural language to shell command translator"
requires-python = ">=3.9"
dependencies = [
    "litellm",
    "rich",
    "pyyaml",
]

[project.scripts]
ai = "ai_shell.cli:entry"

[tool.setuptools.packages.find]
include = ["ai_shell*"]
